feat: Fixed Gateway behavior and output messages
