+++ services/gateway/controllers/stationInfoController.js
@@ -4,9 +4,25 @@ exports.getStationInfo = async (req, res) => {
     try {
         const response = await services.getStationInfo()
 
-        const links = {}
+        const links = {
+            nowPlaying: {
+                href: `${req.protocol}://${req.get("host")}/api/v1/now-playing`,
+                method: "GET",
+                type: "application/json"
+            },
+            listeners: {
+                href: `${req.protocol}://${req.get("host")}/api/v1/listeners`,
+                method: "GET",
+                type: "application/json"
+            },
+            health: {
+                href: `${req.protocol}://${req.get("host")}/api/v1/health`,
+                method: "GET",
+                type: "application/json"
+            }
+        }
 
-        res.hateoas({
+        res.locals.hateoas({
             ...response,
             fetchedAt: new Date().toISOString()
         }, links)
@@ -14,12 +30,12 @@ exports.getStationInfo = async (req, res) => {
     }
     catch (error) {
         console.log(error)
-        res.hateoas({
+        res.locals.hateoas({
             status: 'OFFLINE',
             message: 'Service Unavailable'
         }, {
             retry: { href: req.originalUrl, method: 'GET' },
             status: { href: '/api/v1/health', method: 'GET' }
-        });
+        })
     }
 }
